many warning fixes
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@422 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/VideoCommon/Src/TextureDecoder.cpp
@@ -480,7 +480,7 @@ PC_TexFormat TexDecoder_Decode(u8 *dst, const u8 *src, int width, int height, in
 	while(*fmt)
 	{
 		int xcnt = 0;
-		int nchar = sfont_map[*fmt];
+		int nchar = sfont_map[(int)*fmt];
 		
 		const unsigned char *ptr = sfont_raw[nchar]; // each char is up to 9x10
